requirements.txt
python-telegram-bot
Flask
gunicorn
gspread
oauth2client
gspread-formatting
openpyxl
pandas
